Fix 16-bit→32-bit int adaptation bugs (architecture review §13)
Following the architecture review's Section 13 findings, fix type
mismatches that work coincidentally under sizeof(int)==2 (Pure C,
GCC -mshort) but cause UB or wrong-stride access under modern
m68k-atari-mint-gcc with 32-bit int.

source/window.c (in library):
 - va_arg(short) → va_arg(int). Variadic short args are promoted
   to int per C's default argument promotion; reading them as short
   reads the wrong bytes with 32-bit int.

Sub-application files (not currently built by makefile.gcc, but
fixed for correctness if cross-compiled):
 - viewer/winview.c, viewer/xclipbrd.c, font/font_pal.c,
   font/font_sel.c, rscview/preview.c, monitor/xaccmoni.c,
   play_it/play_it.c: Init/InitEvent/Event/Draw callback
   declarations changed from int to short — Event_Handler and
   USERBLK.ub_code take short-returning function pointers; int
   return is a type mismatch at the function-pointer assignment
   site under 32-bit int.
 - Same files: msg pointers typed short* (aliasing XEVENT.ev_mmgpbuf
   which is short[8]) instead of int*; mbuf[8] short instead of int.
 - viewer/winview.c: *(int*)&ptr[4] → *(short*)&ptr[4] for the
   AV VIEW_DATA protocol header offset field (16-bit by spec).
 - viewer/data.c, monitor/xaccmoni.c: VDI output arrays (out[8],
   out[4]) typed short[] — v_set_text/vqt_extent write 16-bit
   values; 32-bit int slots leave upper halves garbage.
 - viewer/data.h: IMG file-header struct fields short instead of
   int — struct is overlaid on file-buffer bytes (data.c:268) and
   must be 16 bytes per GEM IMG file spec.
 - play_it/play_it.c: three pxy[8] coordinate arrays passed to
   vrt_cpyfm/vro_cpyfm typed short[].

Not addressed:
 - FONTSEL bitfield allocation unit (§13.5): in-program only,
   never serialized or read by offset; layout shift is harmless
   under a single compiler.
 - Implicit function declarations (§13.7): widespread and silenced
   by -Wno-implicit-function-declaration; separate cleanup pass.

+++ source/font/font_pal.c
@@ -526,15 +526,15 @@ int get_type(int obj,int masc)
 	return(ob_isstate(cr_tree,obj,SELECTED) ? masc : 0);
 }
 
-int InitEvent(XEVENT *ev,int avail)
+short InitEvent(XEVENT *ev,short avail)
 {
 	return((MU_MESAG|MU_KEYBD)&avail);
 }
 
-int Event(XEVENT *ev)
+short Event(XEVENT *ev)
 {
 	long ypos;
-	int wi=ev->ev_mwich,*msg=ev->ev_mmgpbuf;
+	short wi=ev->ev_mwich,*msg=ev->ev_mmgpbuf;
 
 	if (wi&MU_MESAG)
 		switch (msg[0])
@@ -745,7 +745,7 @@ void Draw(WIN *win,GRECT *work)
 	restore_clipping(old);
 }
 
-int cdecl DrawUser(PARMBLK *pb)
+short cdecl DrawUser(PARMBLK *pb)
 {
 	GRECT work = *(GRECT *) &pb->pb_x;
 	if (rc_intersect((GRECT *) &pb->pb_xc,&work))
